requirements.txt
websocket-client>=1.9.0
litellm>=1.81.5
